# docs/swagger.yaml
basePath: /api/
definitions:
  domain.Customer:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        type: string
      updated_at:
        type: string
    type: object
  domain.OutgoingCustomer:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  inputs.CreateCustomerRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - email
    - name
    type: object
  inputs.PwdAuth:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  outputs.AuthSuccessResponse:
    properties:
      token:
        type: string
    type: object
  outputs.ErrorResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: A powerful API for managing customers wishlists.
  title: Wishlist API GO
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: user credentials
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/inputs.PwdAuth'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/outputs.AuthSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/outputs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/outputs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/outputs.ErrorResponse'
      summary: Authenticate user
      tags:
      - auth
  /api/customers:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.OutgoingCustomer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/outputs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/outputs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/outputs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Gets given customer data by ID
      tags:
      - customers
    post:
      consumes:
      - application/json
      parameters:
      - description: Client data
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/inputs.CreateCustomerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Customer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/outputs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/outputs.ErrorResponse'
      summary: Creates a new customer
      tags:
      - customers
securityDefinitions:
  BearerAuth:
    in: Header
    name: Authorization
    type: apiKey
swagger: "2.0"